functions sorted, ...

+++ include/arch/syscall.h
@@ -3,13 +3,13 @@
 
 #include "stdint.h"
 
-extern int __syscall0(usize syscallN);
-extern int __syscall1(usize syscallN, usize arg0);
-extern int __syscall2(usize syscallN, usize arg0, usize arg1);
-extern int __syscall3(usize syscallN, usize arg0, usize arg1, usize arg2);
-extern int __syscall4(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3);
-extern int __syscall5(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3, usize arg4);
-extern int __syscall6(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3, usize arg4, usize arg5);
+extern isize __syscall0(usize syscallN);
+extern isize __syscall1(usize syscallN, usize arg0);
+extern isize __syscall2(usize syscallN, usize arg0, usize arg1);
+extern isize __syscall3(usize syscallN, usize arg0, usize arg1, usize arg2);
+extern isize __syscall4(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3);
+extern isize __syscall5(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3, usize arg4);
+extern isize __syscall6(usize syscallN, usize arg0, usize arg1, usize arg2, usize arg3, usize arg4, usize arg5);
 
 #ifdef _LIBC_ARCH_BITS_64
 #include "x86_64/include/_syscallno.h"
